+++ src/core.d.ts
@@ -1,4 +1,0 @@
-export * from './core/client';
-export * from './core/plugin-manager';
-export * from './core/resource-manager';
-export * from './cli';