+++ src/services.d.ts
@@ -1,5 +1,0 @@
-export * from './services/http';
-export * from './services/logger';
-export * from './services/storage';
-export * from './services/updater';
-export * from './services/xmltojson';